requirements.txt
python-dotenv==1.0.0
slack-sdk==3.26.1
aiohttp==3.9.1
beautifulsoup4==4.12.2
playwright==1.40.0
openai==1.3.7
langchain==0.0.350
langchain-openai==0.0.2.post1
pydantic==2.5.2
pydantic-settings==2.1.0
python-jose==3.3.0
requests==2.31.0
tiktoken==0.5.2
typing-extensions==4.8.0
urllib3==2.1.0
fastapi==0.104.1
uvicorn==0.24.0
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
isort==5.12.0
mypy==1.7.1
ruff==0.1.6
